Test cases for getLineNUmberFromJsonLDPath
